Task #827: Inactivity logout.
The server side relies on Devise's Timeoutable feature.  On the client side, a
script sets a timer and uses Ajax calls to the server to keep an eye on the
timeout value (it might change due to activity in another tab or window).

If the session is nearly expired, shows a 30 second countdown banner and gives
you a chance to keep the session alive by clicking a touch-session link.

Devise upgrade needed for the skip_trackable feature to work.